feat: level2 작업 추가
h2 데이터베이스 설정
데이터베이스 스키마 추가
스프링 컨테이너로 객체 관리하도록 변경
daoFactory 삭제
