+++ docs/README.md
@@ -1,5 +1,7 @@
 # 기능요구사항
 
+## 1단계
+
 - [x] 역 dao 수정하기 
 
   - [x] 같은 이름을 갖는 지하철역 생성 불가
@@ -34,6 +36,10 @@
 
   - [x] Factory 이용해서 만들기
   - [x] 싱글턴 패턴 적용하기
-
   
-
+## 2단계
+- [x] gradle 의존성 추가하기
+- [x] h2 데이터베이스 설정
+- [x] h2 스키마 설정
+- [x] 스프링 컨테이너로 객체관리하기
+- [ ] JDBC 활용하는 코드로 변경
