--- src/main/java/wooteco/subway/line/LineDaoFactory.java
@@ -1,12 +1,0 @@
-package wooteco.subway.line;
-
-public class LineDaoFactory {
-
-    private static final LineDao lineDao = new LineDao();
-
-    private LineDaoFactory() {}
-
-    public static LineDao makeLineDao() {
-        return lineDao;
-    }
-}
