+++ src/main/java/wooteco/subway/station/StationController.java
@@ -11,15 +11,22 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import wooteco.subway.station.dto.StationRequest;
+import wooteco.subway.station.dto.StationResponse;
 
 @RestController
 public class StationController {
+
+    private final StationDao stationDao;
+
+    public StationController(StationDao stationDao) {
+        this.stationDao = stationDao;
+    }
 
     @PostMapping("/stations")
     public ResponseEntity<StationResponse> createStation(
         @RequestBody StationRequest stationRequest) {
         Station station = new Station(stationRequest.getName());
-        StationDao stationDao = StationDaoFactory.makeStationDao();
         Station newStation = stationDao.save(station);
         StationResponse stationResponse = new StationResponse(newStation.getId(),
             newStation.getName());
@@ -29,7 +36,6 @@ public class StationController {
 
     @GetMapping(value = "/stations", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<List<StationResponse>> showStations() {
-        StationDao stationDao = StationDaoFactory.makeStationDao();
         List<Station> stations = stationDao.findAll();
         List<StationResponse> stationResponses = stations.stream()
             .map(it -> new StationResponse(it.getId(), it.getName()))
