+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -16,12 +16,19 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.TestConstructor;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.line.dto.LineResponse;
 
 @DisplayName("지하철 노선 관련 기능")
+@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
 public class LineAcceptanceTest extends AcceptanceTest {
 
-    private final LineDao lineDao = LineDaoFactory.makeLineDao();
+    private final LineDao lineDao;
+
+    public LineAcceptanceTest(LineDao lineDao) {
+        this.lineDao = lineDao;
+    }
 
     @BeforeEach
     void beforeSetUp() {
